+++ src/main/java/insane96mcp/survivalreimagined/setup/SREnchantments.java
@@ -1,6 +1,7 @@
 package insane96mcp.survivalreimagined.setup;
 
 import insane96mcp.survivalreimagined.SurvivalReimagined;
+import insane96mcp.survivalreimagined.module.experience.enchantment.MagicProtection;
 import insane96mcp.survivalreimagined.module.experience.enchantment.Magnetic;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraftforge.registries.DeferredRegister;
@@ -11,4 +12,5 @@ public class SREnchantments {
     public static final DeferredRegister<Enchantment> ENCHANTMENTS = DeferredRegister.create(ForgeRegistries.ENCHANTMENTS, SurvivalReimagined.MOD_ID);
 
     public static final RegistryObject<Enchantment> MAGNETIC = ENCHANTMENTS.register("magnetic", Magnetic::new);
+    public static final RegistryObject<Enchantment> MAGIC_PROTECTION = ENCHANTMENTS.register("magic_protection", MagicProtection::new);
 }